requirements.txt
yt-dlp>=2023.03.04
mutagen>=1.46.0
Pillow>=9.0.0